+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.23.0] - 2024-07-08
+
 ### Changed
 
 - Timestepping: the velocity and streamfunction fields are now represented as